requirements.txt
astroid==2.2.5
colorama==0.4.1
Django==2.2.1
django-menu-generator==1.0.4
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
psycopg2==2.8.2
pylint==2.3.1
pytz==2019.1
PyYAML==5.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
wrapt==1.11.1
